GetNonPrimeNos/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.7
django-tastypie==0.14.3
Flask==1.1.2
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
testypie==1.0.0
urllib3==1.26.4
Werkzeug==1.0.1
